remove some predict and metric function from mapsManager
