+++ clinicadl/splitter/split_utils.py
@@ -45,41 +45,24 @@ def find_finished_splits(maps_path: Path, split_name: str) -> List[int]:
     return finished_splits
 
 
-def find_selection_metrics(maps_path: Path, split_name: str, split):
-    """Find which selection metrics are available in MAPS for a given split."""
-
-    split_path = maps_path / f"{split_name}-{split}"
-    if not split_path.is_dir():
-        raise KeyError(
-            f"Training of split {split} was not performed."
-            f"Please execute maps_manager.train(split_list=[{split}])"
-        )
-
-    return [
-        metric.name.split("-")[1]
-        for metric in list(split_path.iterdir())
-        if metric.name.startswith("best-")
-    ]
-
-
-def check_selection_metric(
-    maps_path: Path, split_name: str, split, selection_metric=None
+def print_description_log(
+    maps_path: Path,
+    split_name: str,
+    data_group: str,
+    split: int,
+    selection_metric: str,
 ):
-    """Check that a given selection metric is available for a given split."""
-    available_metrics = find_selection_metrics(maps_path, split_name, split)
+    """
+    Print the description log associated to a prediction or interpretation.
 
-    if not selection_metric:
-        if len(available_metrics) > 1:
-            raise ClinicaDLArgumentError(
-                f"Several metrics are available for split {split}. "
-                f"Please choose which one you want to read among {available_metrics}"
-            )
-        else:
-            selection_metric = available_metrics[0]
-    else:
-        if selection_metric not in available_metrics:
-            raise ClinicaDLArgumentError(
-                f"The metric {selection_metric} is not available."
-                f"Please choose among is the available metrics {available_metrics}."
-            )
-    return selection_metric
+    Args:
+        data_group (str): name of the data group used for the task.
+        split (int): Index of the split used for training.
+        selection_metric (str): Metric used for best weights selection.
+    """
+    log_dir = (
+        maps_path / f"{split_name}-{split}" / f"best-{selection_metric}" / data_group
+    )
+    log_path = log_dir / "description.log"
+    with log_path.open(mode="r") as f:
+        content = f.read()
